Refactor: Update BaseViewModel package
Moved BaseViewModel to the `presentation.screen.base` package.
Updated all import statements across various ViewModels to reflect this change.
