+++ presentation/src/main/java/com/london/presentation/screen/details/actordetails/gallery/ActorGalleryViewModel.kt
@@ -2,9 +2,9 @@ package com.london.presentation.screen.details.actordetails.gallery
 
 import androidx.lifecycle.SavedStateHandle
 import com.london.domain.usecase.GetActorImagesByIdUseCase
-import com.london.presentation.features.base.BaseViewModel
 import com.london.presentation.navigation.Screen
 import com.london.presentation.navigation.getArgs
+import com.london.presentation.screen.base.BaseViewModel
 import org.koin.android.annotation.KoinViewModel
 
 @KoinViewModel
