+++ presentation/src/main/java/com/london/presentation/screen/details/movieDetalis/MovieDetailsViewModel.kt
@@ -3,9 +3,9 @@ package com.london.presentation.screen.details.movieDetalis
 import androidx.lifecycle.SavedStateHandle
 import com.london.domain.usecase.GetMovieDetailsUseCase
 import com.london.domain.usecase.GetMovieVideoUseCase
-import com.london.presentation.features.base.BaseViewModel
 import com.london.presentation.navigation.Screen
 import com.london.presentation.navigation.getArgs
+import com.london.presentation.screen.base.BaseViewModel
 import org.koin.android.annotation.KoinViewModel
 
 @KoinViewModel
