+++ presentation/src/main/java/com/london/presentation/screen/details/tvshow/episodedetails/EpisodeDetailsViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.SavedStateHandle
 import com.london.domain.usecase.GetEpisodeByTvShowId
 import com.london.domain.usecase.GetImagesById
 import com.london.domain.usecase.GetTvShowDetails
-import com.london.presentation.features.base.BaseViewModel
 import com.london.presentation.navigation.Screen
 import com.london.presentation.navigation.getArgs
+import com.london.presentation.screen.base.BaseViewModel
 import org.koin.android.annotation.KoinViewModel
 
 @KoinViewModel
