+++ presentation/src/main/java/com/london/presentation/screen/details/tvshow/tvshowdetails/TvShowDetailsViewModel.kt
@@ -6,9 +6,9 @@ import com.london.domain.usecase.GetEpisodesByTvShowSeason
 import com.london.domain.usecase.GetImagesById
 import com.london.domain.usecase.GetTvShowDetails
 import com.london.domain.usecase.GetTvShowVideoProvider
-import com.london.presentation.features.base.BaseViewModel
 import com.london.presentation.navigation.Screen
 import com.london.presentation.navigation.getArgs
+import com.london.presentation.screen.base.BaseViewModel
 import org.koin.android.annotation.KoinViewModel
 
 @KoinViewModel
